# Supoort more dimensions in forward fast layer_norm kernel