<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Function optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
Supoort cols=3072 for PLAO-XL in forward fast layer_norm kernel.

**Performance**:

| cols | before | after | speedup |
| ---- | ------ | ----- | ------- |
| 3072 | 17.82ms  | 7.10ms  | 2.51x   |

**Previous PR**:
#39167
#42282
#43118